New York City zip codes

A simple features object containing polygons for the 262 zip code areas in New York city, along with additional geographic information.

nyc_zips

Format

A simple feature collection with 262 features (rows) and 11 fields (12 columns).

objectid

Object identifier.

zip_code

Zip code.

po_name

Post office region name.

state

State (all NY).

borough

City borough.

st_fips

State FIPS code (all 36).

cty_fips

County FIPS code.

bld_gpostal_code

Building postal code.

shape_leng

Shape length.

shape_area

Shape area

vstrat

Variance stratum

x_id

x_id

geometry

POLYGON [°] description of the shape of the zip code area.

Source

.