Licensed Dogs in New York City as of January 1st 2017

Dogs in New York city with licenses issued between September 12th 2014 and December 31st 2016.

nyc_license

Format

A tibble containing 122,203 rows and 14 columns

animal_name

Name of dog.

animal_gender

Sex of dog.

animal_birth_month

Birth month.

breed_rc

Breed of dog (simplier categories).

borough

Borough of owner.

zip_code

Zip code of owner.

community_district

Community district of owner

nta

Neighborhood Tabulation Area.

city_council_district

City council district.

congressional_district

Congressional district

state_senatorial_district

State Senatorial district

license_issued_date

Date license issued

license_expired_date

Date license expired

Source

New York City Public Data